Move notes table component
