+++ packages/ui/src/pages/court-cases/[courtCaseId]/reallocate.tsx
@@ -2,7 +2,7 @@ import ActionLink from "components/ActionLink"
 import ConditionalRender from "components/ConditionalRender"
 import { HeaderContainer, HeaderRow } from "components/Header/Header.styles"
 import Layout from "components/Layout"
-import { NotesTable } from "components/NotesTable"
+import { NotesTable } from "features/CourtCaseDetails/Tabs/Panels/Notes/NotesTable"
 import ReallocationNotesForm from "components/ReallocationNotesForm"
 import { CourtCaseContext, useCourtCaseContextState } from "context/CourtCaseContext"
 import { CsrfTokenContext, useCsrfTokenContextState } from "context/CsrfTokenContext"
